Update to EMEW3.
Replace the EMEW2 by EMEW3 format since comma is an invalid character
in message-ids. This causes some MLM to munge the headers because of
invalid characters.

EMEW3|mmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmttttttxxoriginal_sender|original_msg_id

m..m is the MD5 hash of the tttttt, xx, original_sender, and
original_msg_id.

tttttt is the year, month, day, hour, minute, second as a base62
number. The year is modulo 62.

xx is the offset in hex of the at-sign within the original_sender. We
need to know this in case the original_sender contains the EMEW
delimiter within the local- part of the address. Since the EMEW
delimiter can not be a valid domain name character, we can find the
start of the original_msg_id by looking for the next EMEW delimiter
from the at-sign position.

original_sender replaces the at-sign with an EMEW delimter, since the
original_msg_id will already have an at-sign and a message-id can only
have one at-sign.