cmdRcpt() now discards the recipent route if there are no valid recipients.  A forward host should only contain an open connection with valid recipients.
This addresses a special case bug when delivering mail to two
recipients in different domains, where the first one fails a
call-ahead check. The route for the failed recipient was added
to the list of forward hosts without ever being connected. Thus
when the DATA command is forwarded, a bad file descriptor error
is reported and extra DSN generated for an unknown recipient.